Move generators to own file & folder
