+++ src/generators/htmlTextGenerator.ts
